chore: move components to respective folders
