+++ src/app/components/primitives/ProfileDeleteAccount.tsx
@@ -6,7 +6,7 @@ const DeleteAccount = () => (
       <div className="p-6">
         <p className="text-[#5A5A5A] mb-6 font-normal">
           Deleting your account is permanent. All your data, including personal info, and history,
-          will be erased and can't be recovered.
+          will be erased and can&apos;t be recovered.
         </p>
         <button className="text-red-600 border border-red-600 px-6 py-3 rounded-lg hover:bg-red-50 transition-colors font-normal">
           Delete account
